move service functions to the handler file
